+++ app/helpers/admin/claims_helper.rb
@@ -13,10 +13,10 @@ module Admin
     def admin_personal_details(claim)
       [
         [t("admin.teacher_reference_number"), claim.teacher_reference_number],
-        [t("verified_fields.full_name").capitalize, claim.full_name],
-        [t("verified_fields.date_of_birth").capitalize, l(claim.date_of_birth, format: :day_month_year)],
+        [t("govuk_verify_fields.full_name").capitalize, claim.full_name],
+        [t("govuk_verify_fields.date_of_birth").capitalize, l(claim.date_of_birth, format: :day_month_year)],
         [t("admin.national_insurance_number"), claim.national_insurance_number],
-        [t("verified_fields.address").capitalize, sanitize(claim.address("<br>").html_safe, tags: %w[br])],
+        [t("govuk_verify_fields.address").capitalize, sanitize(claim.address("<br>").html_safe, tags: %w[br])],
         [t("admin.email_address"), claim.email_address],
       ]
     end