+++ app/models/claim/permitted_parameters.rb
@@ -17,15 +17,15 @@ class Claim
     private
 
     def claim_attributes
-      Claim::EDITABLE_ATTRIBUTES.dup - verified_claim_attributes
+      Claim::EDITABLE_ATTRIBUTES.dup - claim_attributes_from_govuk_verify
     end
 
     def eligibility_attributes
       {eligibility_attributes: claim.eligibility.class::EDITABLE_ATTRIBUTES.dup}
     end
 
-    def verified_claim_attributes
-      claim.verified_fields.map(&:to_sym)
+    def claim_attributes_from_govuk_verify
+      claim.govuk_verify_fields.map(&:to_sym)
     end
   end
 end