+++ app/models/student_loans/slug_sequence.rb
@@ -51,7 +51,7 @@ module StudentLoans
         sequence.delete("student-loan-how-many-courses") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
         sequence.delete("student-loan-start-date") if claim.no_student_loan? || claim.student_loan_country_with_one_plan?
         sequence.delete("name") if claim.name_verified?
-        sequence.delete("address") if claim.address_verified?
+        sequence.delete("address") if claim.address_from_govuk_verify?
         sequence.delete("date-of-birth") if claim.date_of_birth_verified?
         sequence.delete("gender") if claim.payroll_gender_verified?
         sequence.delete("ineligible") unless claim.eligibility.ineligible?