+++ db/migrate/20200115092652_rename_verified_fields_to_gov_uk_verify_fields.rb
@@ -1,0 +1,5 @@
+class RenameVerifiedFieldsToGovUkVerifyFields < ActiveRecord::Migration[6.0]
+  def change
+    rename_column :claims, :verified_fields, :govuk_verify_fields
+  end
+end